Microservice/requirements.txt
anyio==4.4.0
astrapy==1.3.0
asyncio==3.4.3
blinker==1.8.2
bson==0.5.10
cassandra-driver==3.29.1
cassio==0.1.8
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
exceptiongroup==1.2.1
Flask==3.0.3
geomet==0.2.1.post1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.2
uuid6==2024.1.12
Werkzeug==3.0.3
zipp==3.19.2
